+++ Changes.md
@@ -1,5 +1,6 @@
 v1.2.4(working)
 - Add independent config file.
+- Remove internal server.
 
 v1.2.3
 - Fix bug with bmp file.